feat: update gradle plugin clean tasks
Updates Vaadin Gradle plugin to clean in same way as Maven plugin cleans. Moved  CleanFrontendMojo logic to CleanFrontendUtil in shared flow-plugin-base.
Updated VaadinCleanTask and VaadinBuildFrontendTask.
Added cleanFrontendFiles property to Gradle plugin.

Fixes: #18556
