+++ flow-plugins/flow-gradle-plugin/src/functionalTest/kotlin/com/vaadin/gradle/VaadinSmokeTest.kt
@@ -149,27 +149,13 @@ class VaadinSmokeTest : AbstractGradleTest() {
     @Test
     fun vaadinCleanDoesntDeletePnpmFiles() {
         val pnpmLockYaml = testProject.newFile("pnpm-lock.yaml")
-        val pnpmFileJs = testProject.newFile("pnpmfile.js")
         val pnpmFileCjs = testProject.newFile(".pnpmfile.cjs")
         val webpackConfigJs = testProject.newFile("webpack.config.js")
         testProject.build("vaadinClean")
         expect(false) { pnpmLockYaml.exists() }
-        expect(false) { pnpmFileJs.exists() }
         expect(false) { pnpmFileCjs.exists() }
         // don't delete webpack.config.js: https://github.com/vaadin/vaadin-gradle-plugin/pull/74#discussion_r444457296
         expect(true) { webpackConfigJs.exists() }
-    }
-
-    /**
-     * Tests that VaadinClean task removes TS-related files.
-     */
-    @Test
-    fun vaadinCleanDeletesTsFiles() {
-        val tsconfigJson = testProject.newFile("tsconfig.json")
-        val typesDTs = testProject.newFile("types.d.ts")
-        testProject.build("vaadinClean")
-        expect(false) { tsconfigJson.exists() }
-        expect(false) { typesDTs.exists() }
     }
 
     /**
