+++ flow-plugins/flow-gradle-plugin/src/main/kotlin/com/vaadin/gradle/VaadinFlowPluginExtension.kt
@@ -274,6 +274,8 @@ public abstract class VaadinFlowPluginExtension {
 
     public abstract val reactEnable: Property<Boolean>
 
+    public abstract val cleanFrontendFiles: Property<Boolean>
+
     public fun filterClasspath(@DelegatesTo(value = ClasspathFilter::class, strategy = Closure.DELEGATE_FIRST) block: Closure<*>) {
         block.delegate = classpathFilter
         block.resolveStrategy = Closure.DELEGATE_FIRST
@@ -419,6 +421,8 @@ internal class PluginEffectiveConfiguration(
         .convention(FrontendUtils.isReactRouterRequired(frontendDirectory.get()))
         .overrideWithSystemProperty(InitParameters.REACT_ENABLE)
 
+    var cleanFrontendFiles: Property<Boolean> = extension.cleanFrontendFiles
+            .convention(true)
     /**
      * Finds the value of a boolean property. It searches in gradle and system properties.
      *
@@ -466,7 +470,8 @@ internal class PluginEffectiveConfiguration(
             "skipDevBundleBuild=${skipDevBundleBuild.get()}, " +
             "alwaysExecutePrepareFrontend=${alwaysExecutePrepareFrontend.get()}, " +
             "frontendHotdeploy=${frontendHotdeploy.get()}," +
-            "reactEnable=${reactEnable.get()}" +
+            "reactEnable=${reactEnable.get()}," +
+            "cleanFrontendFiles=${cleanFrontendFiles.get()}" +
             ")"
     companion object {
         internal fun get(project: Project): PluginEffectiveConfiguration =
